Initial resource matching logic
