+++ core-models/build.gradle
@@ -1,3 +1,4 @@
 dependencies {
     compile "org.apache.mesos:mesos:$mesosVersion"
+    testCompile project(':test-utils')
 }
