+++ core-models/src/main/scala/com/mesosphere/usi/core/launching/CommandBuilder.scala
@@ -1,0 +1,9 @@
+package com.mesosphere.usi.core.launching
+
+import org.apache.mesos.v1.{Protos => Mesos}
+
+trait CommandBuilder {
+  def buildCommandInfo(offer: Mesos.Offer, matchedResources: Seq[Mesos.Resource]) = {
+    
+  }
+}
