+++ core-models/src/test/scala/com/mesosphere/usi/core/matching/ScalarResourceRequirementTest.scala
@@ -1,0 +1,24 @@
+package com.mesosphere.usi.core.matching
+
+import com.mesosphere.utils.UnitTest
+import com.mesosphere.usi.core.models.ResourceType
+import com.mesosphere.usi.core.protos.ProtoBuilders
+import org.apache.mesos.v1.{Protos => Mesos}
+import org.scalatest.Inside
+
+class ScalarResourceRequirementTest extends UnitTest with Inside {
+
+  import com.mesosphere.usi.core.protos.ProtoConversions._
+  "ScalarResourceRequirement" should {
+    "match and consume a cpu resource" in {
+      val matchers = ScalarResourceRequirement(ResourceType.CPUS, 2.0)
+
+      val resources = List(ProtoBuilders.newResource(ResourceType.CPUS.name, Mesos.Value.Type.SCALAR, scalar = 10.asProtoScalar))
+      inside(matchers.matchAndConsume(resources)) {
+        case Some(ResourceMatchResult(List(matchedResult), List(unmatched))) =>
+          matchedResult.getScalar.getValue shouldBe 2.0
+          unmatched.getScalar.getValue shouldBe 8.0
+      }
+    }
+  }
+}
