+++ core/src/main/scala/com/mesosphere/usi/core/FakeMesos.scala
@@ -1,13 +1,15 @@
 package com.mesosphere.usi.core
 import akka.NotUsed
 import akka.stream.scaladsl.Flow
+import com.mesosphere.usi.core.protos.ProtoBuilders
 import com.mesosphere.usi.core.models.AgentId
 import org.apache.mesos.v1.{Protos => Mesos}
 import org.apache.mesos.v1.scheduler.Protos.{Call => MesosCall, Event => MesosEvent}
+
 import scala.collection.JavaConverters._
 
 object FakeMesos {
-  import ProtoConversions._
+  import com.mesosphere.usi.core.protos.ProtoConversions._
   import ProtoBuilders._
   val fakeAgentId = AgentId("fake-agent")
   val fakeFrameworkId = ProtoBuilders.newFrameworkId("fake-framework")
